// src/weather/dto/weather.dto.ts
export class WeatherDto {
    time: number;
    temp: number;
    tempMax: number;
    tempMin: number;
    tempSensed: number;
    humidity: number;
    pressure: number;
    wind: number;
    clouds: number;
    rain: number;
    snow: number;
    desc: string;
    short: string;
    timezone: number;
}